xip: the last `UInt64` position, in `loadNext`
The previous pass fixed the two readers and the progress cursor and missed this
one, which is what the compiler stopped on next. Every `handle.offsetInFile` is
now converted at the point of use, and none is assigned raw.

+++ App/NativeToolchain/XipArchive.swift
@@ -293,7 +293,7 @@ enum XipArchive {
             guard let block = try nextBlock(), !block.isEmpty else { return false }
             buffer = block
             position = 0
-            consumed = handle.offsetInFile
+            consumed = Int64(handle.offsetInFile)
             if let progress, contentLength > 0 {
                 progress(min(1, Double(consumed) / Double(contentLength)))
             }
